+++ src/core/index.ts
@@ -1,5 +1,4 @@
 export * from './constants';
-export * from './types';
 export * from './utils';
 export * from './fileWatchers';
 export * from './providers';